requirements.txt
phidata>=2.5.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
pydantic>=2.0.0
python-dateutil>=2.8.2
pytz>=2023.3
requests>=2.31.0
SQLAlchemy>=2.0.0
tqdm>=4.65.0
urllib3>=2.0.0
validators>=0.20.0
openai>=1.0.0
langchain>=0.0.300
python-linkedin-v2>=0.9.4
